# backend/src/main/java/org/rdfarchitect/api/controller/datasets/graphs/shacl/custom/SHACLCustomContentRestController.java
/*
 *    Copyright (c) 2024-2026 SOPTIM AG
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 *
 */

package org.rdfarchitect.api.controller.datasets.graphs.shacl.custom;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.responses.ApiResponse;

import lombok.RequiredArgsConstructor;

import org.apache.jena.riot.RDFFormat;
import org.rdfarchitect.api.controller.Response;
import org.rdfarchitect.api.controller.datasets.graphs.shacl.SHACLFileResponse;
import org.rdfarchitect.database.GraphIdentifier;
import org.rdfarchitect.rdf.graph.source.implementations.GraphFileSourceImpl;
import org.rdfarchitect.rdf.graph.source.implementations.GraphStringSourceImpl;
import org.rdfarchitect.services.ExpandURIUseCase;
import org.rdfarchitect.services.shacl.SHACLExportUseCase;
import org.rdfarchitect.services.shacl.SHACLInsertUseCase;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpHeaders;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import java.nio.charset.StandardCharsets;

@RestController
@RequestMapping("api/datasets/{datasetName}/graphs/{graphURI}/shacl/custom")
@RequiredArgsConstructor
public class SHACLCustomContentRestController {

    private static final Logger logger =
            LoggerFactory.getLogger(SHACLCustomContentRestController.class);

    private final ExpandURIUseCase expandURIUseCase;

    private final SHACLInsertUseCase shaclInsertUseCase;

    private final SHACLExportUseCase shaclExportUseCase;

    @Operation(
            summary = "Replace/Insert shacl",
            description =
                    "Replace or insert a shacl graph stored in a file. Superseded by POST "
                            + "/shacl/documents/file, which adds a file as its own document instead of "
                            + "replacing the graph's default one.",
            tags = {"graph"},
            deprecated = true,
            responses = {@ApiResponse(responseCode = "200")})
    @PutMapping(path = "/file", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
    public String replaceGraphWithFile(
            @Parameter(description = "The name/url of the inquirer.")
                    @RequestHeader(
                            value = HttpHeaders.ORIGIN,
                            required = false,
                            defaultValue = "unknown")
                    String originURL,
            @Parameter(description = "The literal name of the dataset.") @PathVariable
                    String datasetName,
            @Parameter(
                            description =
                                    "The url encoded uri of the graph, or \"default\" to access the default graph.")
                    @PathVariable
                    String graphURI,
            @Parameter(description = "The file containing the shacl graph to be inserted")
                    @RequestParam("file")
                    MultipartFile file) {
        logger.info(
                "Received PUT request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/file\" from \"{}\".",
                datasetName,
                graphURI,
                originURL);

        var extendedGraphURI = expandURIUseCase.expandUri(datasetName, graphURI);

        var graph = new GraphFileSourceImpl(file, extendedGraphURI).graph();
        var graphIdentifier = new GraphIdentifier(datasetName, extendedGraphURI);
        shaclInsertUseCase.replaceCustomSHACLGraph(graphIdentifier, graph);

        logger.info(
                "Sending response to PUT request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/file\" to \"{}\".",
                datasetName,
                graphURI,
                originURL);
        return Response.SUCCESS;
    }

    @Operation(
            summary = "Replace/Insert shacl",
            description =
                    "Replace or insert a shacl graph stored in a Turtle String. Superseded by PUT "
                            + "/shacl/documents/{documentId}, which addresses one document explicitly.",
            tags = {"graph"},
            deprecated = true,
            responses = {@ApiResponse(responseCode = "200")})
    // Raw text, not JSON: Spring reads a String @RequestBody verbatim, so a JSON-quoted
    // body would reach Jena with its surrounding quotes and fail to parse.
    @PutMapping(path = "/string", consumes = MediaType.TEXT_PLAIN_VALUE)
    public String replaceGraphWithGraphString(
            @Parameter(description = "The name/url of the inquirer.")
                    @RequestHeader(
                            value = HttpHeaders.ORIGIN,
                            required = false,
                            defaultValue = "unknown")
                    String originURL,
            @Parameter(description = "The literal name of the dataset.") @PathVariable
                    String datasetName,
            @Parameter(
                            description =
                                    "The url encoded uri of the graph, or \"default\" to access the default graph.")
                    @PathVariable
                    String graphURI,
            @Parameter(description = "The entire SHACL graph as a string") @RequestBody
                    String graphString) {
        logger.info(
                "Received PUT request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/string\" from \"{}\".",
                datasetName,
                graphURI,
                originURL);

        var extendedGraphURI = expandURIUseCase.expandUri(datasetName, graphURI);

        var graph = new GraphStringSourceImpl(graphString, extendedGraphURI).graph();
        var graphIdentifier = new GraphIdentifier(datasetName, extendedGraphURI);
        shaclInsertUseCase.replaceCustomSHACLGraph(graphIdentifier, graph);

        logger.info(
                "Sending response to PUT request: \"/api/datasets/{{}}/graphs/{{}}/shacl/content/string\" to \"{}\".",
                datasetName,
                graphURI,
                originURL);
        return Response.SUCCESS;
    }

    @Operation(
            summary = "export shacl",
            description = "Export the rdf-shacl graph",
            tags = {"shacl"},
            responses = {
                @ApiResponse(
                        responseCode = "200",
                        content = {
                            @Content(mediaType = "text/turtle"),
                            @Content(mediaType = "application/rdf+xml"),
                            @Content(mediaType = "application/rdf+json"),
                            @Content(mediaType = "application/n-triples")
                        })
            })
    @GetMapping("/file")
    public ResponseEntity<byte[]> getCustomSHACLAsFile(
            @Parameter(description = "The requested Datatype.", hidden = true)
                    @RequestHeader("Accept")
                    String acceptHeader,
            @Parameter(description = "The name/url of the inquirer.")
                    @RequestHeader(
                            value = HttpHeaders.ORIGIN,
                            required = false,
                            defaultValue = "unknown")
                    String originURL,
            @Parameter(description = "The literal name of the dataset.") @PathVariable
                    String datasetName,
            @Parameter(
                            description =
                                    "The url encoded uri of the graph, or \"default\" to access the default graph.")
                    @PathVariable
                    String graphURI) {
        logger.info(
                "Received GET request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/file\" from \"{}\".",
                datasetName,
                graphURI,
                originURL);

        var extendedGraphURI = expandURIUseCase.expandUri(datasetName, graphURI);
        var format = SHACLFileResponse.rdfFormat(acceptHeader);

        // fetch data
        var outStream =
                shaclExportUseCase.exportCustomSHACLGraph(
                        new GraphIdentifier(datasetName, extendedGraphURI), format);
        var body = SHACLFileResponse.of(extendedGraphURI, format, outStream);

        logger.info(
                "Sending response to GET request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/file\" to \"{}\".",
                datasetName,
                graphURI,
                originURL);
        return body;
    }

    @Operation(
            summary = "export shacl",
            description = "Export the rdf-shacl graph as String",
            tags = {"shacl"},
            responses = {
                @ApiResponse(
                        responseCode = "200",
                        content = {@Content(mediaType = "text/turtle")})
            })
    @GetMapping("/string")
    public String getCustomSHACLAsString(
            @Parameter(description = "The name/url of the inquirer.")
                    @RequestHeader(
                            value = HttpHeaders.ORIGIN,
                            required = false,
                            defaultValue = "unknown")
                    String originURL,
            @Parameter(description = "The literal name of the dataset.") @PathVariable
                    String datasetName,
            @Parameter(
                            description =
                                    "The url encoded uri of the graph, or \"default\" to access the default graph.")
                    @PathVariable
                    String graphURI) {
        logger.info(
                "Received GET request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/string\" from \"{}\".",
                datasetName,
                graphURI,
                originURL);

        var extendedGraphURI = expandURIUseCase.expandUri(datasetName, graphURI);

        // fetch data
        var shaclString =
                shaclExportUseCase
                        .exportCustomSHACLGraph(
                                new GraphIdentifier(datasetName, extendedGraphURI),
                                RDFFormat.TURTLE)
                        .toString(StandardCharsets.UTF_8);

        logger.info(
                "Sending response to GET request: \"/api/datasets/{{}}/graphs/{{}}/shacl/custom/string\" to \"{}\".",
                datasetName,
                graphURI,
                originURL);
        return shaclString;
    }
}