# frontend/openapi.json
{
...
  "info" : {
    "contact" : {
      "name" : "soptim",
      "url" : "https://www.soptim.de/"
    },
    "description" : "This API provides utilities for editing RDFGraphs that model UML classes using the CIM standard.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "RDFArchitect backend",
    "version" : "0.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/datasets/{datasetName}/graphs/{graphURI}/shacl/custom/file" : {
      "get" : {
        "description" : "Export the rdf-shacl graph",
        "operationId" : "getCustomSHACLAsFile",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/n-triples" : { },
              "application/rdf+json" : { },
              "application/rdf+xml" : { },
              "text/turtle" : { }
            },
            "description" : "OK"
          }
        },
        "summary" : "export shacl",
        "tags" : [
          "shacl"
        ]
      },
      "put" : {
        "deprecated" : true,
        "description" : "Replace or insert a shacl graph stored in a file. Superseded by POST /shacl/documents/file, which adds a file as its own document instead of replacing the graph's default one.",
        "operationId" : "replaceGraphWithFile",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "description" : "The file containing the shacl graph to be inserted",
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [
                  "file"
                ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Replace/Insert shacl",
        "tags" : [
          "graph"
        ]
      }
    },
...
    "/api/datasets/{datasetName}/graphs/{graphURI}/shacl/custom/string" : {
      "get" : {
        "description" : "Export the rdf-shacl graph as String",
        "operationId" : "getCustomSHACLAsString",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "text/turtle" : { }
            },
            "description" : "OK"
          }
        },
        "summary" : "export shacl",
        "tags" : [
          "shacl"
        ]
      },
      "put" : {
        "deprecated" : true,
        "description" : "Replace or insert a shacl graph stored in a Turtle String. Superseded by PUT /shacl/documents/{documentId}, which addresses one document explicitly.",
        "operationId" : "replaceGraphWithGraphString",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "text/plain" : {
              "schema" : {
                "description" : "The entire SHACL graph as a string",
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Replace/Insert shacl",
        "tags" : [
          "graph"
        ]
      }
    },
...
  }
}